+++ sphinxcontrib/osint/plugins/timeline.py
@@ -71,19 +71,28 @@ class Timeline(PluginDirective):
         global get_entries_timelines
         def get_entries_timelines(domain, orgs=None, idents=None, cats=None, countries=None, related=False):
             logger.debug(f"get_entries_timelines {cats} {countries}")
-            return [domain.quest.timelines[e].idx_entry for e in
-                domain.quest.get_timelines(cats=cats, countries=countries)]
+            ret = []
+            for i in domain.quest.get_timelines(cats=cats, countries=countries):
+                try:
+                    ret.append(domain.quest.timelines[i].idx_entry)
+                except Exception as e:
+                    logger.warning(__("Can't get_entries_timelines : %s"), str(e))
+            return ret
         domain.get_entries_timelines = get_entries_timelines
 
         global add_timeline
-        def add_timeline(domain, signature, label, options):
+        def add_timeline(domain, signature, label, node, options):
             """Add a new timeline to the domain."""
             prefix = OSIntTimeline.prefix
             name = f'{prefix}.{signature}'
             logger.debug("add_timeline %s", name)
             anchor = f'{prefix}--{signature}'
             entry = (name, signature, prefix, domain.env.docname, anchor, 0)
-            domain.quest.add_timeline(name, label, idx_entry=entry, **options)
+            try:
+                domain.quest.add_timeline(name, label, idx_entry=entry, **options)
+            except Exception as e:
+                logger.warning(__("Can't add carto %s(%s) : %s"), node["osint_name"], node["docname"], str(e),
+                    location=node)
         domain.add_timeline = add_timeline
 
         global resolve_xref_timeline
@@ -109,7 +118,7 @@ class Timeline(PluginDirective):
                     label = options.pop('label')
                 else:
                     label = osint_name
-                domain.add_timeline(osint_name, label, options)
+                domain.add_timeline(osint_name, label, timeline, options)
                 if env.config.osint_emit_related_warnings:
                     logger.warning(__("TIMELINE entry found: %s"), timeline["osint_name"],
                                    location=timeline)
@@ -143,18 +152,25 @@ class Timeline(PluginDirective):
                 else:
                     alttext = domain.quest.timelines[ f'{OSIntTimeline.prefix}.{timeline_name}'].sdescription
 
-                output_dir = os.path.join(processor.env.app.outdir, '_images')
-                filename = domain.quest.timelines[ f'{OSIntTimeline.prefix}.{timeline_name}'].graph(output_dir)
+                try:
 
-                paragraph = nodes.paragraph('', '')
+                    output_dir = os.path.join(processor.env.app.outdir, '_images')
+                    filename = domain.quest.timelines[ f'{OSIntTimeline.prefix}.{timeline_name}'].graph(output_dir)
 
-                image_node = nodes.image()
-                image_node['uri'] = f'/_images/{filename}'
-                image_node['candidates'] = '?'
-                image_node['alt'] = alttext
-                paragraph += image_node
+                    paragraph = nodes.paragraph('', '')
 
-                container.append(paragraph)
+                    image_node = nodes.image()
+                    image_node['uri'] = f'/_images/{filename}'
+                    image_node['candidates'] = '?'
+                    image_node['alt'] = alttext
+                    paragraph += image_node
+
+                    container.append(paragraph)
+
+                except Exception as e:
+                    logger.warning(__("Can't create timeline %s : %s"), node["osint_name"], str(e),
+                               location=node)
+
                 node.replace_self(container)
 
         processor.process_timeline = process_timeline